seminar-3 hot fixes
