+++ practice/practise-3/src/test/java/hse/kpo/factories/PedalCarFactoryTests.java~
@@ -1,0 +1,32 @@
+package hse.kpo.factories;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import hse.kpo.params.PedalEngineParams;
+import hse.kpo.services.CarService;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+
+/**
+ * Pedal car factory tests.
+ */
+public class PedalCarFactoryTests {
+
+    @Mock
+    private PedalCarFactory mockPedalCarFactory;
+
+    @InjectMocks
+    private CarService mockCarService;
+
+    @Test
+    @DisplayName("Adding car using DI")
+    void testAddCar() {
+        int initialSize = mockCarService.getCarNumberCounter();
+        mockCarService.addCar(mockPedalCarFactory, new PedalEngineParams(2));
+        assertEquals(initialSize + 1, mockCarService.getCarNumberCounter(),
+                "Car count should increase by 1");
+    }
+
+}
